+++ backend/src/back_api/get_localion.py
@@ -16,8 +16,8 @@ async def get_loc(
     data: GetLocationRequest,
     geo: Annotated[GeoService, Depends(get_geo_service)],
 ) -> GetLocationResponse:
-    latitude = geo.dms_to_dd(data.degrees_n, data.minutes_n, data.seconds_n)
-    longitude = geo.dms_to_dd(data.degrees_e, data.minutes_e, data.seconds_e)
+    latitude = geo.dms_to_degrees(data.degrees_n, data.minutes_n, data.seconds_n)
+    longitude = geo.dms_to_degrees(data.degrees_e, data.minutes_e, data.seconds_e)
     location = geo.get_location_names(latitude, longitude)
     return GetLocationResponse(
         country=location.country,
