+++ backend/src/back_api/get_records_file.py
@@ -8,8 +8,8 @@ from sqlalchemy.ext.asyncio import AsyncSession
 
 from back_api.rate_limiter import limiter
 from database.database import get_session
-from service.record import RecordExcelService, get_record_excel_service
-from service.record_service import RecordService, get_record_service
+from service.export import ExportService, get_record_excel_service
+from service.record import RecordService, get_record_service
 from service.token import get_current_user
 
 logger = logging.getLogger(__name__)
@@ -23,7 +23,7 @@ async def get_records_data(
     user_data: Annotated[dict, Depends(get_current_user)],
     session: Annotated[AsyncSession, Depends(get_session)],
     records_svc: Annotated[RecordService, Depends(get_record_service)],
-    records_excel: Annotated[RecordExcelService, Depends(get_record_excel_service)],
+    records_excel: Annotated[ExportService, Depends(get_record_excel_service)],
 ) -> StreamingResponse:
     user_id = int(user_data["sub"])
     username = user_data["username"]
@@ -43,7 +43,7 @@ async def get_records_data(
         }
 
         return StreamingResponse(
-            records_excel.generate_excel(records),
+            records_excel.records_to_excel(records),
             media_type="application/vnd.openxmlformats-officedocument.spreadsheetml.sheet",
             headers=headers,
         )
