+++ backend/src/back_api/users.py
@@ -9,7 +9,7 @@ from back_api.schemas import Message, UserRequest
 from config.config import ACCESS_TOKEN_EXPIRE, REFRESH_TOKEN_EXPIRE
 from database.database import get_session
 from service.token import TokenService, get_token_service
-from service.user_service import UserService, get_user_service
+from service.user import UserService, get_user_service
 
 logger = logging.getLogger(__name__)
 router = APIRouter()
@@ -25,7 +25,7 @@ async def handle_user_data(
     tokens: Annotated[TokenService, Depends(get_token_service)],
     users: Annotated[UserService, Depends(get_user_service)],
 ) -> Message:
-    user_id = await users.get_user_by_username(session, data.username)
+    user_id = await users.get_by_username(session, data.username)
     if user_id is None:
         logger.warning("User not found for this username")
         raise HTTPException(status_code=404, detail="User not found for this username")
