+++ backend/src/service/publication.py
@@ -10,22 +10,20 @@ logger = logging.getLogger(__name__)
 
 
 class PublicationService:
-    async def get_publication_by_id(
-        self, session: AsyncSession, publ_id: int
-    ) -> Publ | None:
+    async def get(self, session: AsyncSession, publ_id: int) -> Publ | None:
         return await publication_repo.get_publication(session, publ_id)
 
-    async def get_publications(
+    async def get_for_language(
         self, session: AsyncSession, language: str
     ) -> Sequence[int]:
         return await publication_repo.get_publications_for_language(session, language)
 
-    async def is_filled(
+    async def is_filled_by_user(
         self, session: AsyncSession, user_id: int, publ_id: int
     ) -> bool:
         return await publication_repo.is_publ_filled(session, user_id, publ_id)
 
-    async def add_publication(self, session: AsyncSession, publ_json: dict) -> None:
+    async def add(self, session: AsyncSession, publ_json: dict) -> None:
         await publication_repo.add_publication_from_json(session, publ_json)
