+++ backend/src/service/record.py
@@ -1,92 +1,53 @@
-import io
 import logging
 from collections.abc import Sequence
 
-from openpyxl import Workbook
-from openpyxl.styles import Font
-from openpyxl.utils import get_column_letter
-from starlette.responses import ContentStream
+from sqlalchemy.ext.asyncio import AsyncSession
 
 from database.models import Record
+from model import PublData
+from repository import record as record_repo
 
 logger = logging.getLogger(__name__)
-COLUMN_MAPPING = {
-    "datetime": "Дата добавления записи",
-    "adm_country": "Страна",
-    "adm_region": "Регион",
-    "adm_district": "Район",
-    "adm_loc": "Место сбора",
-    "geo_nn": "Широта (десятич.)",
-    "geo_ee": "Долгота (десятич.)",
-    "geo_nn_raw": "Широта (изнач.)",
-    "geo_ee_raw": "Долгота (изнач.)",
-    "geo_origin": "Происхождение координат",
-    "geo_REM": "Примечания к расположению",
-    "eve_YY": "Год",
-    "eve_MM": "Месяц",
-    "eve_DD": "День",
-    "eve_day_def": "Определён ли день",
-    "eve_habitat": "Биотоп",
-    "eve_effort": "Выборочное усиление",
-    "abu_coll": "Коллектор",
-    "eve_REM": "Примечания к сбору материала",
-    "tax_fam": "Семейство",
-    "tax_gen": "Род",
-    "tax_sp": "Вид",
-    "tax_sp_def": "Определён ли вид",
-    "tax_nsp": "Описан ли как новый вид",
-    "type_status": "Типовой статус",
-    "tax_REM": "Таксономические примечания",
-    "abu": "Общее кол-во особей",
-    "abu_details": "Кол-во особей каждого пола/зрелости",
-    "abu_ind_rem": "Комментарий к особям",
-    "geo_uncert": "Радиус неточности координат, м",
-    "eve_YY_end": "Конечный год",
-    "eve_MM_end": "Конечный месяц",
-    "eve_DD_end": "Конечный день",
-}
 
 
-class RecordExcelService:
-    def generate_excel(self, records: Sequence[Record]) -> ContentStream:
-        output = io.BytesIO()
-        wb = Workbook()
-        ws = wb.active
+class RecordService:
+    async def add(self, session: AsyncSession, record_json: dict) -> None:
+        await record_repo.add_record_from_json(session, record_json)
 
-        if ws is None:
-            logger.error("ws is None")
-            raise Exception
+    async def get_stats(self, session: AsyncSession) -> dict:
+        return await record_repo.get_statistics(session)
 
-        headers = [
-            COLUMN_MAPPING[field]
-            for field in COLUMN_MAPPING
-            if field not in ["id", "publ_id", "ip", "errors", "type", "adm_verbatim"]
-        ]
-        ws.append(headers)
+    async def get_by_user(
+        self, session: AsyncSession, user_id: int
+    ) -> Sequence[Record]:
+        return await record_repo.get_user_records(session, user_id)
 
-        for col in range(1, len(headers) + 1):
-            ws.column_dimensions[get_column_letter(col)].width = 20
-            ws.cell(row=1, column=col).font = Font(bold=True)
+    async def delete(self, session: AsyncSession, record_id: int, user_id: int) -> bool:
+        return await record_repo.remove_record_row_by_id(session, record_id, user_id)
 
-        for record in records:
-            row = [
-                getattr(record, field)
-                for field in COLUMN_MAPPING
-                if field
-                not in ["id", "publ_id", "ip", "errors", "type", "adm_verbatim"]
-            ]
-            ws.append(row)
+    async def get(
+        self, session: AsyncSession, record_id: int, user_id: int
+    ) -> Record | None:
+        return await record_repo.get_record_by_id(session, record_id, user_id)
 
-        wb.save(output)
-        output.seek(0)
-        yield output.read()
+    async def update(
+        self, session: AsyncSession, record_id: int, user_id: int, new_data: dict
+    ) -> bool:
+        return await record_repo.edit_record_by_id(
+            session, record_id, user_id, new_data
+        )
+
+    async def get_publication_by_hash(
+        self, session: AsyncSession, record_id: int, user_id: int
+    ) -> PublData | None:
+        return await record_repo.publ_by_hash(session, record_id, user_id)
 
 
-_record_excel_service: RecordExcelService | None = None
+_record_service: RecordService | None = None
 
 
-def get_record_excel_service() -> RecordExcelService:
-    global _record_excel_service
-    if _record_excel_service is None:
-        _record_excel_service = RecordExcelService()
-    return _record_excel_service
+def get_record_service() -> RecordService:
+    global _record_service
+    if _record_service is None:
+        _record_service = RecordService()
+    return _record_service
